change to uint16

+++ sequential/include/ChinesePostmanProblem.hpp
@@ -9,20 +9,20 @@ class ChinesePostmanProblem
 {
 private:
     Multigraph *mg;
-    void listPairs(vector<uint32_t> &oddVertices, vector<vector<uint64_t>> &distances, vector<map<uint32_t, list<uint32_t>>> &paths);
-    vector<vector<pair<uint32_t, uint32_t>>> listPairsCombinations(vector<uint32_t> &oddVertices);
-    vector<pair<uint32_t, uint32_t>> bestPairsCombination(vector<vector<pair<uint32_t, uint32_t>>> &pairCombinations, vector<vector<uint64_t>> &distances);
-    void modifyGraph(vector<pair<uint32_t, uint32_t>> &bestPairs, vector<map<uint32_t, list<uint32_t>>> &paths);
+    void listPairs(vector<uint16_t> &oddVertices, vector<vector<uint16_t>> &distances, vector<map<uint16_t, list<uint16_t>>> &paths);
+    vector<vector<pair<uint16_t, uint16_t>>> listPairsCombinations(vector<uint16_t> &oddVertices);
+    vector<pair<uint16_t, uint16_t>> bestPairsCombination(vector<vector<pair<uint16_t, uint16_t>>> &pairCombinations, vector<vector<uint16_t>> &distances);
+    void modifyGraph(vector<pair<uint16_t, uint16_t>> &bestPairs, vector<map<uint16_t, list<uint16_t>>> &paths);
 
-    vector<pair<uint32_t, uint32_t>> listPairsCombinationsBase(vector<uint32_t> &oddVertices, vector<vector<uint64_t>> &distances);
-    uint64_t distancePairCombination(vector<pair<uint32_t, uint32_t>> &pairCombinations, vector<vector<uint64_t>> &distances);
+    vector<pair<uint16_t, uint16_t>> listPairsCombinationsBase(vector<uint16_t> &oddVertices, vector<vector<uint16_t>> &distances);
+    uint16_t distancePairCombination(vector<pair<uint16_t, uint16_t>> &pairCombinations, vector<vector<uint16_t>> &distances);
     
 
 public:
     ChinesePostmanProblem(/* args */);
     ~ChinesePostmanProblem();
-    void solve(Multigraph *mg, uint32_t startVertex);
-    void solve_v2(Multigraph *mg, uint32_t startVertex);
+    void solve(Multigraph *mg, uint16_t startVertex);
+    void solve_v2(Multigraph *mg, uint16_t startVertex);
 };
 
 #endif
